Refactor remove contact function to helper file
