+++ src/components/manageContacts/ManageContacts.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import firebase from './../../firebase';
-import fetchScopedUsers from './../../utils/fetchScopedUsers';
+import {fetchScopedUsers, removeContact} from './../../utils/fetchScopedUsers';
 
 class ManageContacts extends Component {
   constructor() {
@@ -14,6 +14,7 @@ class ManageContacts extends Component {
     this.handleChange = this.handleChange.bind(this);
     this.handleSubmit = this.handleSubmit.bind(this);
     this.fetchScopedUsers = fetchScopedUsers.bind(this);
+    this.removeContact = removeContact.bind(this);
   }
 
   componentDidMount() {
@@ -49,23 +50,7 @@ class ManageContacts extends Component {
   }
 
   handleRemove(contact) {
-    const contactsFromDB = firebase.database().ref('contacts');
-
-    contactsFromDB.on('value', snapshot => {
-      const contactEntries = Object.entries(snapshot.val());
-
-      const removeContact = contactEntries.map(contact => {
-        return contact[0];
-      }).filter(contactId => {
-        return contactId === contact.id;
-      });
-      console.log(removeContact[0]);
-      const contactRef = firebase.database().ref(`contacts/${removeContact[0]}`);
-      console.log(contactRef);
-      contactRef.remove();
-      //this.props.removeFromFB(removeContact);
-    })
-//this deleted ALL the contacts.
+    this.removeContact(contact, firebase)
   }
 
   render() {
