+++ src/utils/fetchScopedUsers.js
@@ -1,4 +1,4 @@
-export default function fetchScopedUsers(firebase) {
+export function fetchScopedUsers(firebase) {
   const contactsFromDB = firebase.database().ref('contacts');
 
   contactsFromDB.on('value', snapshot => {
@@ -15,3 +15,21 @@ export default function fetchScopedUsers(firebase) {
     this.props.loadContacts(contacts);
   });
 };
+
+export const removeContact = (contact, firebase) => {
+  const contactsFromDB = firebase.database().ref('contacts');
+
+  contactsFromDB.on('value', snapshot => {
+    const contactEntries = Object.entries(snapshot.val());
+
+    const removeContact = contactEntries.map(contact => {
+      return contact[0];
+    }).filter(contactId => {
+      return contactId === contact.id;
+    });
+
+    const contactRef = firebase.database().ref(`contacts/${removeContact[0]}`);
+
+    contactRef.remove();
+  });
+}
